requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
addict=2.4.0=pypi_0
ansys-additive-core=0.17.0=pypi_0
ansys-api-additive=1.4.1=pypi_0
ansys-api-dbu=0.2.2=pypi_0
ansys-api-dyna=0.3.5=pypi_0
ansys-api-fluent=0.3.20=pypi_0
ansys-api-geometry=0.3.5=pypi_0
ansys-api-mapdl=0.5.1=pypi_0
ansys-api-mechanical=0.1.1=pypi_0
ansys-api-meshing-prime=0.1.2=pypi_0
ansys-api-platform-instancemanagement=1.0.0=pypi_0
ansys-api-pyensight=0.3.4=pypi_0
ansys-api-sherlock=0.1.22=pypi_0
ansys-api-systemcoupling=0.1.0=pypi_0
ansys-dpf-composites=0.3.2=pypi_0
ansys-dpf-core=0.10.1=pypi_0
ansys-dpf-gate=0.4.1=pypi_0
ansys-dpf-gatebin=0.4.1=pypi_0
ansys-dpf-post=0.6.0=pypi_0
ansys-dyna-core=0.4.5=pypi_0
ansys-dynamicreporting-core=0.5.1=pypi_0
ansys-fluent-core=0.18.2=pypi_0
ansys-geometry-core=0.4.6=pypi_0
ansys-grantami-bomanalytics=2.0.0=pypi_0
ansys-grantami-bomanalytics-openapi=2.0.0=pypi_0
ansys-grantami-recordlists=1.1.0=pypi_0
ansys-grantami-serverapi-openapi=2.0.0=pypi_0
ansys-grpc-dpf=0.8.1=pypi_0
ansys-mapdl-core=0.67.0=pypi_0
ansys-mapdl-reader=0.53.0=pypi_0
ansys-math-core=0.1.3=pypi_0
ansys-mechanical-core=0.10.6=pypi_0
ansys-mechanical-env=0.1.2=pypi_0
ansys-meshing-prime=0.5.0=pypi_0
ansys-motorcad-core=0.3.0=pypi_0
ansys-openapi-common=1.4.0=pypi_0
ansys-optislang-core=0.5.1=pypi_0
ansys-platform-instancemanagement=1.1.2=pypi_0
ansys-pyensight-core=0.7.5=pypi_0
ansys-pythonnet=3.1.0rc3=pypi_0
ansys-seascape=0.2.0=pypi_0
ansys-sherlock-core=0.4.0=pypi_0
ansys-systemcoupling-core=0.3.0=pypi_0
ansys-tools-path=0.4.1=pypi_0
ansys-turbogrid-api=0.4.dev4=pypi_0
ansys-turbogrid-core=0.4.dev1=pypi_0
appdirs=1.4.4=pypi_0
apptools=5.2.1=pypi_0
asgiref=3.7.2=pypi_0
asttokens=2.4.1=pypi_0
attrs=23.2.0=pypi_0
bcrypt=4.1.2=pypi_0
beartype=0.16.4=pypi_0
bleach=6.1.0=pypi_0
blinker=1.7.0=pypi_0
bokeh=3.3.3=pypi_0
build=1.0.3=pypi_0
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.12.12=h06a4308_0
cachetools=5.3.2=pypi_0
certifi=2023.11.17=pypi_0
cffi=1.16.0=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
clr-loader=0.2.6=pypi_0
comm=0.2.1=pypi_0
configargparse=1.7=pypi_0
configobj=5.0.8=pypi_0
contourpy=1.2.0=pypi_0
cryptography=41.0.7=pypi_0
cycler=0.12.1=pypi_0
cyrus-sasl=2.1.28=h52b45da_1
dash=2.15.0=pypi_0
dash-core-components=2.0.0=pypi_0
dash-html-components=2.0.0=pypi_0
dash-table=5.0.0=pypi_0
dbus=1.13.18=hb2f20db_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
deprecated=1.2.14=pypi_0
dill=0.3.7=pypi_0
distro=1.9.0=pypi_0
django=5.0.1=pypi_0
docker=7.0.0=pypi_0
docutils=0.20.1=pypi_0
dotnetcore2=3.1.23=pypi_0
elementpath=4.1.5=pypi_0
envisage=7.0.3=pypi_0
exceptiongroup=1.2.0=pypi_0
executing=2.0.1=pypi_0
expat=2.5.0=h6a678d5_0
fabric=3.2.2=pypi_0
fake-bpy-module=20240222=pypi_0
fastjsonschema=2.19.1=pypi_0
filelock=3.13.1=pypi_0
flask=3.0.2=pypi_0
fontconfig=2.14.1=h4c34cd2_2
fonttools=4.48.1=pypi_0
fpdf2=2.7.7=pypi_0
freetype=2.12.1=h4a9f257_0
geomdl=5.3.1=pypi_0
glib=2.69.1=he621ea3_2
google-api-core=2.15.0=pypi_0
google-api-python-client=2.114.0=pypi_0
google-auth=2.26.2=pypi_0
google-auth-httplib2=0.2.0=pypi_0
googleapis-common-protos=1.62.0=pypi_0
grpcio=1.60.0=pypi_0
grpcio-health-checking=1.48.2=pypi_0
grpcio-status=1.48.2=pypi_0
gst-plugins-base=1.14.1=h6a678d5_1
gstreamer=1.14.1=h5eee18b_1
httplib2=0.22.0=pypi_0
icu=73.1=h6a678d5_0
idna=3.6=pypi_0
importlib-metadata=6.11.0=pypi_0
invoke=2.2.0=pypi_0
ipython=8.21.0=pypi_0
ipywidgets=8.1.1=pypi_0
itsdangerous=2.1.2=pypi_0
jaraco-classes=3.3.0=pypi_0
jedi=0.19.1=pypi_0
jeepney=0.8.0=pypi_0
jinja2=3.1.3=pypi_0
joblib=1.3.2=pypi_0
jpeg=9e=h5eee18b_1
jsonschema=4.21.1=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
jupyter-core=5.7.1=pypi_0
jupyterlab-widgets=3.0.9=pypi_0
keyring=24.3.0=pypi_0
kiwisolver=1.4.5=pypi_0
krb5=1.20.1=h143b758_1
ld_impl_linux-64=2.38=h1181459_1
libclang=14.0.6=default_hc6dbbc7_1
libclang13=14.0.6=default_he11475f_1
libcups=2.4.2=h2d74bed_1
libedit=3.1.20230828=h5eee18b_0
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libllvm14=14.0.6=hdb19cb5_3
libpng=1.6.39=h5eee18b_0
libpq=12.17=hdbd6064_0
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.1=h5eee18b_1
libxml2=2.10.4=hf1b16e4_1
linkify-it-py=2.0.2=pypi_0
lxml=5.1.0=pypi_0
lz4-c=1.9.4=h6a678d5_0
markdown=3.5.2=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.3=pypi_0
matplotlib=3.8.2=pypi_0
matplotlib-inline=0.1.6=pypi_0
mayavi=4.8.1=pypi_0
mdit-py-plugins=0.4.0=pypi_0
mdurl=0.1.2=pypi_0
more-itertools=10.2.0=pypi_0
mysql=5.7.24=h721c034_2
nbformat=5.9.2=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pypi_0
nh3=0.2.15=pypi_0
numpy=1.26.4=pypi_0
open3d=0.18.0=pypi_0
openssl=3.0.13=h7f8727e_0
packaging=23.2=pypi_0
pandas=2.1.4=pypi_0
panel=1.3.7=pypi_0
param=2.0.2=pypi_0
paramiko=3.4.0=pypi_0
parso=0.8.3=pypi_0
pcre=8.45=h295c915_0
pexpect=4.9.0=pypi_0
pillow=10.2.0=pypi_0
pint=0.23=pypi_0
pip=23.3.1=py310h06a4308_0
pkginfo=1.9.6=pypi_0
platformdirs=4.1.0=pypi_0
plotly=5.18.0=pypi_0
plumbum=1.8.2=pypi_0
ply=3.11=py310h06a4308_0
pooch=1.8.0=pypi_0
prompt-toolkit=3.0.43=pypi_0
protobuf=3.20.3=pypi_0
psutil=5.9.8=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pyaedt=0.7.10=pypi_0
pyansys=2024.1.3=pypi_0
pyansys-tools-versioning=0.5.0=pypi_0
pyasn1=0.5.1=pypi_0
pyasn1-modules=0.3.0=pypi_0
pycparser=2.21=pypi_0
pyface=8.0.0=pypi_0
pygments=2.17.2=pypi_0
pygranta=2024.1.0=pypi_0
pyiges=0.3.1=pypi_0
pynacl=1.5.0=pypi_0
pyparsing=3.1.1=pypi_0
pypng=0.20220715.0=pypi_0
pyproject-hooks=1.0.0=pypi_0
pyqt=5.15.10=py310h6a678d5_0
pyqt5-sip=12.13.0=py310h5eee18b_0
pyquaternion=0.9.9=pypi_0
pyspnego=0.10.2=pypi_0
python=3.10.13=h955ad1f_0
python-dateutil=2.8.2=pypi_0
pytwin=0.6.0=pypi_0
pytz=2023.3.post1=pypi_0
pyvista=0.43.2=pypi_0
pyviz-comms=3.0.1=pypi_0
pyyaml=6.0.1=pypi_0
qt-main=5.15.2=h53bd1ea_10
readline=8.2=h5eee18b_0
readme-renderer=42.0=pypi_0
referencing=0.33.0=pypi_0
requests=2.31.0=pypi_0
requests-ntlm=1.2.0=pypi_0
requests-toolbelt=1.0.0=pypi_0
retrying=1.3.4=pypi_0
rfc3986=2.0.0=pypi_0
rich=13.7.0=pypi_0
rpds-py=0.17.1=pypi_0
rpyc=5.3.1=pypi_0
rsa=4.9=pypi_0
scikit-learn=1.4.0=pypi_0
scipy=1.11.4=pypi_0
scooby=0.9.2=pypi_0
secretstorage=3.3.3=pypi_0
setuptools=68.2.2=py310h06a4308_0
sip=6.7.12=py310h6a678d5_0
six=1.16.0=pypi_0
sqlite=3.41.2=h5eee18b_0
sqlparse=0.4.4=pypi_0
stack-data=0.6.3=pypi_0
tenacity=8.2.3=pypi_0
threadpoolctl=3.2.0=pypi_0
tk=8.6.12=h1ccaba5_0
tomli=2.0.1=py310h06a4308_0
tornado=6.4=pypi_0
tqdm=4.66.1=pypi_0
traitlets=5.14.1=pypi_0
traits=6.4.3=pypi_0
traitsui=8.0.0=pypi_0
twine=4.0.2=pypi_0
typing-extensions=4.9.0=pypi_0
tzdata=2023.4=pypi_0
uc-micro-py=1.0.2=pypi_0
uritemplate=4.1.1=pypi_0
urllib3=1.26.18=pypi_0
vtk=9.3.0=pypi_0
wcwidth=0.2.13=pypi_0
webencodings=0.5.1=pypi_0
werkzeug=3.0.1=pypi_0
wheel=0.41.2=py310h06a4308_0
widgetsnbextension=4.0.9=pypi_0
wrapt=1.16.0=pypi_0
xmlschema=2.5.1=pypi_0
xyzservices=2023.10.1=pypi_0
xz=5.4.5=h5eee18b_0
zipp=3.17.0=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
